requirements.txt
pandas
numpy
scikit-learn
matplotlib
python-dotenv
PyYAML
